Remove Gemfile.lock
Having the Gemfile.lock checked into the gem doesn't make sense because
we want to test the gem against the dependencies it resolves to, not the
dependencies we lock it to. When it's installed in another ruby project
it would resolve it's dependencies and not use what we lock it to here.